+++ src/api/resources/post_air_quality_response.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class PostAirQualityResponse(BaseModel):
+    air_quality: float
