+++ src/api/resources/runway.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class Runway(BaseModel):
+    length: float
+    surface: str
